lesson19/python_databases/requirements.txt
mysql-connector-python==8.0.22
psycopg2-binary==2.8.6
sqlalchemy==1.3.20